move component files to subfolder and update references
