+++ resources/assets/js/components/layout/Header.vue
@@ -33,7 +33,7 @@
           @click.stop="onRestartButtonClick"
           class="header-nav__restart-button"
         >
-          <img src="./assets/restart.svg" />
+          <img src="../assets/restart.svg" />
           <span>Restart</span>
         </div>
