+++ resources/assets/js/components/messages/CarouselListMessage.vue
