+++ resources/assets/js/components/messages/FpRichMessage.vue
