+++ resources/assets/js/components/messages/HandToHumanMessage.vue
