+++ resources/assets/js/components/messages/ImageMessage.vue
