+++ resources/assets/js/components/messages/LongTextMessage.vue
