+++ resources/assets/js/components/messages/RichMessage.vue
