+++ resources/assets/js/components/messages/TextMessage.vue
