+++ resources/assets/js/components/messages/inputs/FullPageRichInput.vue
